- Models: added option for AdamW optimizer to `compile_model()` in utils_models.py. - Models: test_model_architecture.py is now more flexible to test different models in model_keras.py and saves model summary and plot
