+++ models/test_model_architecture.py
@@ -6,32 +6,34 @@ Used to test building models.
 import yaml
 from tensorflow.keras.utils import plot_model
 
-
 # local
-from models.models_keras import ExoMinerPlusPlus, ExoMinerJointLocalFlux, ExoMinerMLP
 from src.utils.utils_dataio import set_tf_data_type_for_features
+from models import models_keras
 
 # load file with features and model config
-yaml_config_fp = '/nobackupp19/msaragoc/work_dir/Kepler-TESS_exoplanet/codebase/tess_spoc_ffi/config_cv_train_ffi.yaml'
+yaml_config_fp = '/nobackupp19/msaragoc/work_dir/Kepler-TESS_exoplanet/codebase/models/model_config.yaml'
+model_fn = 'model_exominer_tess-spoc-ffi_unfolded_flux_joint_5-8-2025_1311'
+model_plot_fp = f'/u/msaragoc/work_dir/Kepler-TESS_exoplanet/experiments/test_exominer_architectures/{model_fn}_plot.png'
+model_summary_fp = f'/u/msaragoc/work_dir/Kepler-TESS_exoplanet/experiments/test_exominer_architectures/{model_fn}_summary.txt'
 
 with open(yaml_config_fp, 'r') as file:
     config = yaml.unsafe_load(file)
 
-# load model hyperparameters from HPO run; overwrites the one in the yaml file
-# hpo_dir = Path('/nobackupp19/msaragoc/work_dir/Kepler-TESS_exoplanet/experiments/hpo_configs/hpo_merged_unfolded_7-5-2023')
-# config_hpo_chosen, config['hpo_config_id'] = load_hpo_config(hpo_dir)
-# config['config'].update(config_hpo_chosen)
-
 config['features_set'] = set_tf_data_type_for_features(config['features_set'])
 
-# model = ExoMinerPlusPlus(config, config['features_set']).kerasModel
-model = ExoMinerJointLocalFlux(config, config['features_set']).kerasModel
-# model = ExoMinerMLP(config, config['features_set']).kerasModel
+base_model = getattr(models_keras, config['model_architecture'])
 
+model = base_model(config, config['features_set']).kerasModel
+
+# get model summary
+with open(model_summary_fp, 'w') as f:
+    model.summary(print_fn=lambda x: f.write(x + '\n'))
+
+# plot model architecture
 plot_model(model,
-           to_file='/home6/msaragoc/model_exominer_new_novar.png',
-           show_shapes=True,
-           show_layer_names=True,
-           rankdir='TB',
-           expand_nested=False,
-           dpi=48)
+   to_file=model_plot_fp,
+   show_shapes=True,
+   show_layer_names=True,
+   rankdir='TB',
+   expand_nested=False,
+   dpi=48)
